\name{PenHMM-package}
\alias{PenHMM-package}
\alias{PenHMM}
\title{Penalized Hidden Markov Models with Covariates}
\description{
The \pkg{PenHMM} package provides functions to simulate, estimate, and cross-validate hidden Markov models with covariates in the measurement model using a penalized maximum likelihood approach.
}
\details{
Main functions are \code{Draw.HMCovManifest}, \code{Estimate.HMCovManifest}, and \code{Select.CV}.
}
\keyword{package}
